Remove copies of lars.py and byol_pytorch.py.
